core/requirements.txt
Django==4.0.5
Pillow==9.1.1
django-bootstrap-form==3.4
django-filter==21.1
requests==2.26.0
django-allauth==0.48.0